balance changes for Drill and Hammers

+++ src/main/java/pt/omegaleo/survivalessentials/util/tools/HammerTool.java
@@ -6,6 +6,7 @@ import javax.lang.model.util.ElementScanner6;
 
 import net.minecraft.block.Block;
 import net.minecraft.block.BlockState;
+import net.minecraft.block.Blocks;
 import net.minecraft.entity.LivingEntity;
 import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.item.IItemTier;
@@ -60,17 +61,12 @@ public class HammerTool extends PickaxeItem
                 BlockPos[] blocksToDestroy = getAOEPositions(pos, mop.getFace());
                 for(int i = 0; i < blocksToDestroy.length; i++)
                 {
-                    if(worldIn.isBlockPresent(blocksToDestroy[i]))
+                    if(worldIn.isBlockPresent(blocksToDestroy[i]) && getBlock(blocksToDestroy[i], worldIn) != Blocks.BEDROCK)
                     {
                         worldIn.destroyBlock(blocksToDestroy[i], true);
                     }
-                }
-                    
-                int damage = damagePerUse;
-    
-                if((stack.getMaxDamage() - stack.getDamage()) - damagePerUse <= 0)
-                {
-                    damage = stack.getMaxDamage() - stack.getDamage() - 1;
+
+                    stack.setDamage(stack.getDamage() + 1);
                 }
             }
         }
@@ -79,6 +75,12 @@ public class HammerTool extends PickaxeItem
             System.out.println(e.getStackTrace());
         }
         return super.onBlockDestroyed(stack, worldIn, state, pos, entityLiving);
+    }
+
+    private Block getBlock(BlockPos pos, World world) {
+        BlockState ibs = world.getBlockState(pos);
+        Block block = ibs.getBlock();
+        return block;
     }
